Make exceptions public
